requirements.txt
flask
gunicorn
tensorflow
pillow
gdown